requirements.txt
streamlit
pandas
numpy
requests
geopy
streamlit-folium
folium
scikit-learn
scipy
Pillow